fix eval fred and skip eval to float

+++ configs.py
@@ -28,7 +28,7 @@ def parse_arguments():
     parser.add_argument("--lora_rank", type=int, default=8)
     parser.add_argument("--lora_alpha", type=int, default=32)
     parser.add_argument("--lora_dropout", type=float, default=0.1)
-    parser.add_argument("--skip_eval_ep", type=int, default=0, help="Skip evaluation for the first n epochs")
+    parser.add_argument("--skip_eval_ep", type=float, default=0, help="Skip evaluation for the first n epochs")
     parser.add_argument("--task_ep", type=int, default=10, help="Number of epochs for each task")
     parser.add_argument("--use_mole", action="store_true")
     parser.add_argument("--mole_num_experts", type=int, default=4)
@@ -73,7 +73,7 @@ def parse_arguments():
     parser.add_argument('--task_num', default=5, type=int)
     parser.add_argument('--early_stop', action='store_true')
     parser.add_argument('--patience', type=int, default=2)
-    parser.add_argument('--eval_freq', type=int, default=5)
+    parser.add_argument('--eval_freq', type=float, default=5)
 
     parser.add_argument('--input_map', action='store_true', help="Whether to use input mapping, if False, use span_s to predict trigger type")
     parser.add_argument('--class_num', type=int, default=10)
